+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ChangeLog
 ==========
 
-v1.1
+v1.2 (July 7, 2014)
+----
+- FIXED : Bugged reset highscore button
+- FIXED : Enemy AI improved
+- Customized buttons
+- Added the option to mute the background music in start and pause menu
+- Enemy speed increases as score increases to increase difficulty
+
+v1.1 (July 6, 2014)
 ----
 - FIXED : Bugged diagonal movement of player which gave player more speed than moving straight
 - Improved Graphics; player sprite, enemy sprite and food sprite